+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Removed
+- swish, use `torch.nn.functional.silu` instead
+
 ## [0.2.2] - 2021-02-09
 ### Changed
 - Euler angle convention from ZYZ to YXY